requirements.txt
customtkinter
Pillow
opencv-python
requests
ffpyplayer